프로필 이미지 기능 및, 파일 검사 기능 구현중

+++ src/main/java/com/live/main/profile/database/dto/ProfileImageDto.java
@@ -1,6 +1,5 @@
 package com.live.main.profile.database.dto;
 
-import com.live.main.user.database.entity.UsersEntity;
 import lombok.Data;
 
 @Data
@@ -8,7 +7,7 @@ public class ProfileImageDto {
   private Long id;
   private String imageName;
   private String imageUrl;
-  private int size;
+  private Long size;
   private String fileType;
   private boolean isUser;
   private Long userId;
